feat: #28/JPA와 Querydsl 상속 제거
