+++ src/main/java/com/sparta/delivery_app/domain/liked/repository/UserReviewLikedRepository.java
@@ -3,6 +3,6 @@ package com.sparta.delivery_app.domain.liked.repository;
 import com.sparta.delivery_app.domain.liked.entity.UserReviewLiked;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserReviewLikedRepository extends JpaRepository<UserReviewLiked, Long>, UserReviewLikedQueryRepository{
+public interface UserReviewLikedRepository extends JpaRepository<UserReviewLiked, Long> {
 
 }
