+++ src/main/java/com/sparta/delivery_app/domain/user/repository/UserRepository.java
@@ -1,13 +1,12 @@
 package com.sparta.delivery_app.domain.user.repository;
 
 import com.sparta.delivery_app.domain.user.entity.User;
+import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long>, UserQueryRepository {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
